+++ packages/types/src/material.ts
@@ -64,7 +64,7 @@ export type PropertyGroup = {
 }
 
 export type PropertyWidget = {
-  component: 'input' | 'color' | 'select' | 'switch' | 'slider' | 'upload' | 'json-editor';
+  component: 'input' | 'number' | 'color' | 'select' | 'switch' | 'slider' | 'upload' | 'json-editor';
   inline?: boolean, // 是否行内显示
   props: Record<string, any>;
 }