+++ packages/designer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ruomu-ui/designer
 
+## 1.0.2
+
+### Patch Changes
+
+- @ruomu-ui/core@1.0.0
+- @ruomu-ui/plugin-material@2.0.2
+
 ## 1.0.1
 
 ### Patch Changes