+++ packages/plugins/material/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ruomu-ui/plugin-material
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ruomu-ui/api@1.0.2
+  - @ruomu-ui/core@1.0.0
+
 ## 2.0.1
 
 ### Patch Changes